split and refactor symbol search & delete components
