+++ frontend/src/components/SymbolsListItem.js
@@ -1,21 +1,16 @@
-import { useEffect, useState } from "react";
-import { useNavigate } from "react-router-dom";
+import Box from '@mui/material/Box';
+import Card from '@mui/material/Card';
+import CardActions from '@mui/material/CardActions';
+import CardContent from '@mui/material/CardContent';
+import CircularProgress from '@mui/material/CircularProgress';
+import Tooltip from '@mui/material/Tooltip';
+import Typography from '@mui/material/Typography';
+import { useEffect, useState } from 'react';
+import { useNavigate } from 'react-router-dom';
+import store from 'store-js';
 
-import Box from "@mui/material/Box";
-import Button from "@mui/material/Button";
-import Card from "@mui/material/Card";
-import CardActions from "@mui/material/CardActions";
-import CardContent from "@mui/material/CardContent";
-import CircularProgress from "@mui/material/CircularProgress";
-import Grid from "@mui/material/Grid";
-import Tooltip from "@mui/material/Tooltip";
-import Typography from "@mui/material/Typography";
-
-import ForwardIcon from "@mui/icons-material/ArrowForward";
-
-import store from "store-js";
-
-import CollapsedSection from "./CollapsedSection";
+import CollapsedSection from './CollapsedSection';
+import ContinueButton from './ContinueButton';
 
 const SymbolsListItem = ({ name, filterValue }) => {
   const navigate = useNavigate();
@@ -88,18 +83,7 @@ const SymbolsListItem = ({ name, filterValue }) => {
         </CollapsedSection>
       </CardContent>
       <CardActions>
-        <Grid container justifyContent="flex-end">
-          <Button
-            size="small"
-            startIcon={<ForwardIcon />}
-            onClick={() => {
-              store.set("selected_symbols", symbols);
-              navigate("/process");
-            }}
-          >
-            Continue with these symbols
-          </Button>
-        </Grid>
+        <ContinueButton symbols={symbols} />
       </CardActions>
     </Card>
   );
