+++ frontend/src/utils/helpers.js
@@ -1,4 +1,6 @@
-import dayjs from "dayjs";
+import dayjs from 'dayjs';
+
+import { getSymbols } from '../api';
 
 export const shortenString = (word, length) => {
   return word.length > length ? `${word.slice(0, length)}...` : word;
@@ -6,4 +8,16 @@ export const shortenString = (word, length) => {
 
 export const formatDate = (date) => {
   return dayjs(date, "YYYY-MM-DD").format("DD MMM YYYY");
+};
+export const filterUniqueSymbols = (symbols) => {
+  return [
+    ...new Map(symbols.map((symbol) => [symbol["symbol"], symbol])).values(),
+  ];
+};
+
+export const fetchAllSymbols = async () => {
+  const symbols = await getSymbols();
+  return filterUniqueSymbols(symbols).sort((a, b) =>
+    a.symbol > b.symbol ? 1 : -1
+  );
 };
